Allow for fractional baudrates (#59)
This PR resolves calculating prescaler when a fractional baud rate is
requested (IE: 83333bps or 666666bps)

The prescaler function errors out if `f_can/f_out*tq` does not equal 0.
This means for fractional baudrates, the MCAN's clock would also need to
be fractional, which is very hard to achieve on most MCUs.

Therefore, this PR adjusts the prescaler function to allow for a maximum
error of 0.05% (CAN has a tolerance of 0.75% on CAN FD and ~1% on normal
CAN), rather than requiring a perfect clock combination.

Closes: #58